requirements.txt
numpy
torch
tqdm
sentence-transformers
umap-learn
datamapplot
matplotlib
transformers
pillow
scikit-learn
python-dotenv
datasets
pyyaml
einops
jaxtyping
opencv-python
wandb
